refactor: moved internal business logic to internals directory
