+++ packages/rxtree/src/internals/bestSplit.js
